robotino-clips-agent: prototype of CLIPS-based agent
This plugin uses the CLIPS production (rule-based) system to create a
series of actions to perform in the Festo logistics league. The current
version only simulates execution of behaviors.

It works (so far)! It created a first P today after just a few hours
from four initial products (the minimum required).

Needs more testing and adjustments, but it's a base to start from.