robotino-agent: added skiller support
Added code to actually call skills on the skiller. Currently uses an
extended simulation that calls out to the C++ part, which will wait a
defined time (instead of actually calling the skiller) and then declares
success. The sim.clp is still used for now. This allows to test durative
actions and CLIPS/C++ interaction.