skills:goto: change param naming (see details)
The parameter goto_name is now disabled and leads to failure since the
machine_pos_module is defunct. Instead, we support a param place which
is forwarded to ppgoto. The param names goto_x, goto_y and goto_ori
still work, but using x, y, ori is now supported and actually preferred.