llsf2012: modified 2012 agent to play against the 2013 refbox
This is an intermediate version that is able to play the old game
against the (slightly modified) refbox of the 2013 game. This required a
few changes on the internal representation to be more in-sync with
things we did and learned for the refbox CLIPS code. Adding the current
version of the refbox messages. Along the way fixed a bug or two in the
code to be able to complete a game. This agent can play autonomously
against the refbox and score the ideal amount of 130 points (as long as
it does not "over-produce" P1s in a particular phase).