motor_move: be friggin' awesome
tf_module was fixed to use get_yaw, so the branch vmatare/get_yaw_tolua
NEEDS TO BE MERGED in the fawkes repository. Otherwise, the skiller will
fail in strange and frightening ways.

On the bright side, the formerly horrible and messed-up motor_move is
now small, super-fast, super-tunable, efficient and even a little
documented (see code comments).