Puck_vision: added missing config parameter
You can now use either the Similarity color model or the classic
colormap for puck detection by entering the suitable string in the
puck_vision.yaml config.

For Calibrating, take a screenshot of the camera picture,
get the RGB of a puck by using the Pipette, and put the values
in puck_vision.yaml at the reference_color section.