agent2014: exploration strategy lines + nearest
There are three horizontal lines, one for each robot.
Each robot explores all machines of his line starting in our half of the
field. Machines of the other team will be skipped.
After finishing the line, the robot explores the nearest machines
beginning with the nearest machine to the end of the line.