puck_vision: puck detection enhanced
Search for RED ROIS, than search again in the lowest boundary
of the subroi ror RED. This will seperate pucks which are on
the same distance. To verify that we have a puck, this ROI
gets enlarged for the contain also the yellow dots.
Than the yellow dots are searched, if more than 2 are found its a
puck.