deliver_puck: make use of the new machine detection
Now the robot can track all three gates from one
position, and drive to the open one. Then it checks
again if it is still open and performs the delivery.
If it's closed right before the deliver, the robot
will move back to the starting point and start the whole
procedure again.

The skill is in heavy testing and will be cleaned up later.