agent-monitor: plugin for monitoring an llsf agent
This plugin should provide an easy possibility to monitor the LLSF-agent
with its current state, intention etc in the webview.
Later the resulting page should be included in a new webview tab that
monitors all three LLSF agents.
Currently, the plugin only shows the factbase as it is done by clips-webview.

Conflicts:
	src/plugins/Makefile