skills:deliver_puck: use new machine_signal delivery
Delivery is now done by looking at two signals at once (instead of all
three). The machine_signal plugin returns the position of the open
delivery gate directly in a Position3DInterface, which is then handed
down to move_under_rfid. This is expected to accelerate the delivery
behavior significantly.