navgraph-generator-2014: proof of concept navgraph generator
This plugin uses the Voronoi navgraph generator to create a basic graph
and then augments it with special points of interest, i.e. machine
positions. It also automatically generates the appropriate properties
for the nodes.

The nodes of interest are copied from the existing graph, as are the
default properties. For the 2015 game, we will need to compute these
points based on recognized machine positions.

This plugin serves as a proof of concept. A new plugin is to be written
which actually provides an interface for the reasoning component to
instruct addition of obstacles and points of interest to trigger
continuous updates of the graph (throughout the exploration phase).