ax12_gripper: plugin to use Robotis servos as a gripper
This plugin is based on the pantilt-plugin from the fawkes-core-repository
and it is used to control Robotis Dynamixel AX12 Servos used as a gripper.
The plugin offers an interface of type AX12GripperInterface which lets the
user open, close and preferably close the gripper until a specified load
is given.