ax12_gripper: add skill to use with the ax12_gripper
The skill is based on the gripper.lua skill found in the
fawkes-robotino-repository and lets the user interact with the
ax12_gripper plugin by making use of the AX12GripperInterface.
The combination of the interface and the skill lets the user close the
gripper until a defined maximum load is given on each servo resulting in
the particular servo stopping its motion.
This lets the gripper try to grab a puck which isn't straight in front
of the gripper but lays apart of it in a small offset.