ax12_gripper: change max_torque policy from plugin to servo
Watching the max load of a servo via interface and stopping on max load
can lead to the problem that the sensor thread doesn't get called in
time thus resulting in a not always working solution for the
stop-on-max-load functionality.
To handle this problem, the max_load handling is now controlled by the
max_torque functionality implemented in the servo.