ax12_gripper: adaption to dynamixel-plugin
The dynamixel-plugin from the fawkes-core-repository now handles
communication with dynamixel servos and serves a
DynamixelServoInterface.
The ax12_gripper_thread makes use of this interface and keeps the
earlier functionality of the ax12_gripper-plugin by using the two servos
(defined in ax12gripper.yaml) with a device-specific prefix (for
example /dynamixel/ax12gripper) to identify the desired dynamixel-bus
from the DynamixelServoInterface.
The full path for a given servo is then for example
"/dynamixel/ax12gripper/1" for the servo with ID 1.

The ax12_gripper-plugin doesn't communicate on its own with servos any
longer but via messages to the DynamixelServoInterface for each servo.