ax12_gripper: add plugin-wise goto_gripper until load
The AX12ServoInterface serves a CloseLoad-message offering a
functionality to close the gripper until a given load is reached for
each servo.
The plugin monitors the given load on the left and right servos by using
the DynamixelServoInterface until __cfg_max_load * 1023
(__cfg_max_load is given in a range from 0 to 1) is reached and sends
a StopMessage via the DynamixelServoInterface for the particular servo to
stop motion.