gazsim-machine-signal: simulate light detection of machine-signal
This plugin gets the ground truth information about the light signal of
an mps in front of the robot. It is produces similar results
to gazsim-light-front but the computation on which signal the robot is
looking at the moment is done in the gazebo-plugin and there are no more
special cases about delivery-lights. I did not change the old plugin, so
that the 2014 game can still be simulated.

Conflicts:
	cfg/conf.d/meta_plugins.yaml