navgraph-prefer-exploration: increase cost of non-exploration edges
This is a domain-specific plugin which increases the cost of all
non-exploration edges for as long as there are any exploration edges.
This is to prefer the grid structure during exploration while traveling.
Once all zones have been marked explored and no more exploration edges
are copied from the base graph no more cost factors are applied.