ax12_gripper: add center-message to interface
The center-message lets the user move both servos in a centered position
by keeping the given opening angle.
Note that this function temporarily disables the prevent_alarm_shutdown
functionality of the dynamixel-plugin by sending the
SetPreventAlarmShutdown-Message to the dynamixel-plugin. This is needed
to make sure that both servos get a chance to reach the desired
position.