ax12_gripper: process new ax12_gripper holds_puck
The plugin now correctly sets the interface-value holds_puck depending
on the given load measured by both servos.
It sums up the loads of both servos and the angles of both servos.
When the load is higher than the configuration-wise given threshold
(load_for_holds_puck_threshold) and the opening angle is higher than the
matching config-value (angle_for_holds_puck_threshold), it sets the
interface to true.