explore_zone: add laser-cluster
add a cluster to check if anything is within the area to expolre.
For this a new laser-cluster was added to detect any obj.

Also the skill explore_zone drives with drive_to_global to the zone
first before it drives with drive_to_local between the different point
within a zone.