agent: generic worldmodel synchronization
The worldmodel synchronization is now independent from the concrete
deftemplates. To enable synchronization for a deftemplate the template
needs the slot sync-id and has to be listed in the (wm-sync-info) fact
as synchronized template. Therefore, changing or adding new deftemplates
does not requre more effort for synchronization or rebuilding the
protobuf messages.

To change something locally and synchronize the change with the
synchronized bots the (synced-modify) function can be used. It is used
as the (modify) function but without brackets inside. Similar, there are
(synced-assert), (synced-retract), (synced-add-to-multifield), and
(synced-remove-from-multifield).

The protobuf-messages for the synchronization use a key-value structure.
The key is composed of the sync-id of the corresponding fact and the
slot-number. There are special keys for asserting and retracting.
The worldmodel messages from the master are sets of those key-value msgs
and split so that each worldmodel message stays under a certain size.