conveyor_vision: setup houghlines calculation
The houghlines are now calculated by making use of binarization, canny
edge detection, gap closing, masking unneeded areas (by drawing black
rectangles over never-used areas to concentrate on line detection and
further remove noise) of the image gathered by the cascade classifier.
This includes the dependency of the houghline detection to get a ROI
defined by the cascade classifier.

From the set of detected houghlines the two nearest lines from the middle are
selected in each frame. Then a configurable average of these lines is
calculated to get equalized results.
A resulting line showing the length of the belt is drawn in the
shared memory buffer for debugging reasons.

Note: The houghline detection only delivers a better estimate for the
distance from camera to the conveyor belt. Y- and Z- values are not
improved (yet).