skills/goto: periodically re-calc relgoto params
So far, the relative goto parameters were calculated only once at the
start of the skill. If the pose estimate at that time was bad, this
might lead to a wrong position. Additionally, the odometry error
accumulated over time must be corrected.

Therefore we now re-calculate the relative position (based on the
robot's current pose estimate) once per second.