agent: use id to solve two masters conflict
When there are two masters detected, only the agent with the higher id
becomes Slave. This is faster than the old approach (both become Slave
and wait a random time before trying to become master again), because
there is no time without a master and the conflic can not appear a
second time by drawing similar timeouts.