webtools-bridge: Working Version [Dispatcher/mapping] not the final
Here the Client_socket is in the ros_proxy instance owned by the
Dispatcher.  Yet the client handeing is in the dispatcher and called on
the client_socket property in the rosproxy.

To Be Done:
Seperate the client in a class which will have the: Socket,
write_operation

the ros_proxy and fawkes_proxy extending a parent class. The templetes
should impelment process_req(client_msg) and deals with the
client.