webtools-bridge: Streams Manipulation
After removing lots of cofusions about the concepts of stream..The
ostringstream that was used to process the streambuf (which has the
received msgs) was replaced with an std::istream.

The rapidjson can only read UTF-8 byte streams directly..Since i am not
sure which encoding is arriving from the sockect i want to make it
flixable. Luckly Rapidjson has AutoUTFInputStream that will detect
encoding. Problem is you have to pass a CustomStream implementing their
peek4() function that helps with the detection..

For that i created a istream wrapper to pass it to the
AutoUTFInputStream