webtools-bridge: Json extraction: Stream is okay
This is a json encoding format felixable version. But the problem is not
the byte encoding at all as it tunrs out.I have been trying to pasre a
websocket fram not a json msg. so the parsing part should be fine but
before that I have to decode the fram header and extract the json msg.
Problem is i am not sure if that is possible because it looks like
websockt encrypts it..Let see