webtools-bridge: [Adding Websocketpp]
So turns out i made a big mistake trying to decode the client msg
directly and thinking it was a json.. A big detail I have missed is that
the msg coming in a websocket frame Payload.. Which is even Masked to
make things worst!

So to peek into the msg I can either keep the current structure and try
to decode it myself. Which is gonna be very complected and unnecessary.

Found an ASIO based, header only Websocket Library that I will use to
either to only get the payload..Or to even re-structure the whole Proxy
and make a none transparent proxy.

If I did that I  have to make a change of terminology even. So The
websockt SERVER will be what the web client connects to. and the full
web socket connection will be intialzied between My Plugin and the
webclinet....The plugin should simulate a rosBridge connection form this
side. From another side (the rosBridge) it would be fine if it stayed
the same ever.