webtools-bridge: Dispatcher initializing the ros_client from dispatcher
The thread initiated in the endpoint to run the transport in the
background,  was always terminated if i copied the endpoint_shared_ptr to
a global shared_ptr of the same type. even though the thread keeps
running if i only exited the init() without trying to copy the
shared_ptr.

it worked When passed the global shared_ptr to the function as a parameter and let it do the binding