webtools-bridge: Dispatcher intializing the ros_client from dispatcher
The thread intiated in the endpoint to run the transport in the
backgound,  was always terminate if i copied the endpoint_shared_ptr to
a global shared_ptr of the same type. even though the thread keeps
running if i only exited the init() without trying to copy the
shared_ptr.

it wokred When passed the global shared_ptr to the function as a paramter and let it do the binding