webtools-bridge: Dispatcher refactoring (shared_ptr problem resolved)
Last commit the client could start the Dispatcher and its run thread.
The dispatcher in turn initiats ros_endpoint and its run thread. without
blocking ..But when the web_client thread calls on_message segmentation
fault happens when it tries to access the shared_ptr of the
ros_endpoint.

It is caused by a thread to access an object that is not thread safe.
Solution was to initialized all the pointers before running if the
threads.