webtools-bridge: segmentation bug located and fixed
Whenever the we try to access the connection_ptr from web_sessoin
( through endpiont_->get_connection_from_hdl(hdl_) or while sending to the web_session )
a segmentation falut happens. This did not happen when the transport part was running independantly from the bridge. 2 things in my mind could be the reason. Either the connection_ptr is not thread safe or the wed_session itself is the problem. Anyways this was avoided as a tmp solution To have the server register the dispatchers_msg_hndlr (avoiding any interaction with the session)
but for the forward case it still problem