webtools-bridge: Session_Bug(fix) Report and big restructuring Motivation
Regeneration:
Crashes whenever a second session closes. You can open as many sessions as you want. You can close the first session. But when any other session closes it crashes with back trace shown at the end.

Bug locating and reason:
The implementation of the terminate_callbacks registration and diregistration at the WebSession. The WebSession stores the terminate_callbacks registered by different capabilities as a list of Boost:functions (a result of Boost:binding with  Capability's Session_terminate_hdl method). The problem shows when trying to  deregister_callback. Deleting an element of the boost::function list, uniquely identified by the function's owning Object, not really possible. when deleting using a pointer to member function, the deregistered will delete an element that matches the type. Which is every boost::function that is registered for same Capability type.(the need to use a pointer to member arise since one can't compare boost::function , as far as my search went http://www.boost.org/doc/libs/1_59_0/doc/html/function/faq.html#idp211887664 )

The reason why it worked for the 1st session's closing but not starting the second, is that the difference in the calling order when the of the registration/de registration operation when the topic is new or when it already exists. When a new session is added to the a subscription the subscription the subscription termination_h dler is registered at the session side. If the topic is new u keep this callback at the top of ur list. But when u subscribe and the processor creates a temp Subscription Object. This object will be subsumed later on and when u try to unregistered it from the session the wrong boost::function will be removed

temp Fix / permanent Fix :
TempFix Adding the request (including the session registration) if applied only to the "Active" Subscription Object (i,e. Only on the object that u will keep at the end) avoids this problem all together since u wont need to de register a temporary Subscription object. in short, Just insure that the session only ever receives callbacks registration from subscriptions that will stay and one per topic. this fix only avoids the problem but it does not elemenate the bug. U still can not delete a unique Callback from the boost::functions list

permanent Fix: the whole implementation of howto keep track of the callbacks should garantee to find a unique object's handler. One idea to do that is to actually store the subscription Objects themselves (as  pure virtual type implementing the terminate_handler) as key in a map

conclusiuon:
This bug mad me realize how messy things are and all the levels of indirection that could be avoided with the proper little resurrecting of the Architecture. Hence, I will not waist time in fixing this bug because ambiguity is inherited in this design..I will solve the problem from the core trying to get as much levels of indirection thrown away. The structure I arrived at the solves all these problems (in my opinion) and yet preserves decoupling will follow up with explanation

BackTrace:

    at /home/mgomaa/fawkes-robotino/fawkes/src/libs/core/threading/mutex.cpp:92
    initially_lock=<optimized out>)
    at /home/mgomaa/fawkes-robotino/fawkes/src/libs/core/threading/mutex_locker.cpp:109
    std::shared_ptr (count 7, weak 1) 0x7fffe00240e8)
    at /home/mgomaa/fawkes-robotino/src/plugins/webtools-bridge/subscription_capability.cpp:272
    at /usr/include/boost/bind/mem_fn_template.hpp:165
    at /usr/include/boost/bind/bind.hpp:313
    at /usr/include/boost/bind/bind_template.hpp:32
    at /usr/include/boost/function/function_template.hpp:153
    at /usr/include/boost/function/function_template.hpp:767
    at /home/mgomaa/fawkes-robotino/src/plugins/webtools-bridge/web_session.cpp:102
    hdl=std::weak_ptr (count 14, weak 9) 0x7fffe0017d30)
    at /home/mgomaa/fawkes-robotino/src/plugins/webtools-bridge/Web_server.cpp:108
    this=<optimized out>) at /usr/include/c++/4.8.3/functional:601
    at /usr/include/c++/4.8.3/functional:1296
    at /usr/include/c++/4.8.3/functional:1355
    at /usr/include/c++/4.8.3/functional:2471
    this=0x7fffe0017d30, tstat=<optimized out>, ec=...)
    at /usr/include/websocketpp/impl/connection_impl.hpp:1540
    shutdown_timer=std::shared_ptr (count 4, weak 0) 0x7fffe0024b88, callback=..., ec=...)
    at /usr/include/websocketpp/transport/asio/connection.hpp:1100