agent: worldmodel-sync skip modifying unchanged multislots
CLIPS needed a long time for synchronizing the worldmodel on a slave
because for every multislot in fact in the worldmodel the fact was
modified. Now this is skipped when the value of the multislot was not
changed. This reduces the time needed for the synchronization at least
by a quarter, probably much more.