mps_align: kinda-working implementation
mps_align now has a concept of "interesting lines", i.e. laser lines of
which we don't know whether they might have the tag we're looking for.
It also attempts to track laser lines to weed out lines that have
already been inspected. Since this tracking is unreliable, the "correct"
line may get lost in the process, so the tracking state is reset when we
run out of interesting lines and turn around freely to generate new
hypotheses.