agent: abort loading 4th base into RS
The problem of a 4th base in the RS slide arised from a race condition
of a previous bot removing first that it loads the slide and afterwards
incrementing the base counter. In between another bot could decide to
load an additional base. This decision is now aborted when the counter
is incremented.