arduino: add read_pending
The read_pending flag is supposed to indicate whether we're waiting for
the arduino to send something or not.
The z_movement_pending flag determines whether we sent a movement
command before and are waiting for the arduino to send the Idle-message.
This way we can set the interface-flag "Final" regarding the current
arduino's status.