arduino: set current_z_position
The current_z_position is directly gathered from the values stored in
the AccelStepper-Instance on the arduino.
Everytime the Arduino sends an Idle-package it also sends the current
steps-position. This value is divided by the number of steps per mm and
stored in the arduino-interface to get the current position in mm.