agent: locking for machine sides and modularized base insertion
The machine locking mechanism was only suited locking whole machines.
Often machines can be used from both sides at the same time, depending
on the tasks. Now only input or output of a machine is being locked,
keeping the other side free for another agent.

In a first step the mounting of caps to finalize products has been
modularized. Instead of calling bring_product_to from anywhere on the
field the robot now drives to the insertion side of the machine first,
before calling the skill. This allows for a better control of the game
in regard to machine states.