navgraph-basis: default conveyor-laser line y offset for better mps_align
To properly see the conveyor with the new conveyor vision we need to
roughly align in front of the conveyor. The y offset between conveyor
and the machine center (laser line) is defined in the navgraph and given
to mps_align. The conveyor should be mounted by around 2.5 cm from the
machine center.