approach_mps: skill for approaching using the laser sensor
To make the mps approaching more robust before picking or placing a
product, this skill uses the distance to the mps as determined by the
laser_front_dist plugin. You call the skill with the parameter x for the
desired distance, and the skill moves to that distance independent of
the initial distance. Except the distance is too large or the
visibility_history is <= 0 (e.g. if there are 'nan' laser beams).