agent: add additional ring with waiting for RS to avoid deadlock
We had a deadlock caused by a C3 with already one ring mounted in RS1
and a Bot using RS1 when the product is removed and the second ring also
needing to be mounted in RS1.
This deadlock is now solved by having an additional production rule with
low priority to get the C3, waits at a waiting point until the RS can be used,
 and then mounts the next ring. This made the new step 'wait-for-rs'
necessary that drives to the waiting point, and finished when the next
ring can be mounted and the lock for the RS is aquired.