agent: lock BS before instruction and release after get-base
Because we want to make sure that the instruction was processed by the
BS we want to wait for it to be ready. Furthermore when standing with
multiple bots at the BS we have to lock the instruction and get-base
step to be sure which instruction is processed first and which side is
used first by which bot.