ASP-Planer: Reworked the grounding and solving. (Again.)
There is now the possibility to restart the complete process, as advised
by R. Kaminski. Not used yet, because we need to store which information
should be added again after resetting.

While we still need exploration this needs to be saved. Apart from that
each begin, update and end predicate should be regrounded. But not all,
the begins for explore tasks are invalid when we start all over again.