ASP-Planer: Updated config.
Stop spamming (deactivate debug), increase look ahaed to some useful
value.

Activate more models, because clingo doesn't deliver the optimal model
first, in terms of plans the first model often doesn't contain any
task to do and by that is not usable.
With this in mind the configure variable can be removed, because there
is no situation where you want the false case.