ASP-Planer: New location handling for ASP.
Do not assign the driving durations for each robot to the locations.
This is not useful at all, because we do not consider that the robot is
moving in our plan. Instead assign the durations between the locations
and only put the robots location into the program when the robot is
(re-)added.