ASP-Planer: Merged explorationTransition into transition again.
We will reset the complete clingo environment, when we have found all
machines and get rid off the then unneeded externals etc.
We won't load the exploration.lp in that case and so it can be in the
same program part again, this makes it easier on the C++ part.