ASP-Planer: Improve getBase task.
We have R bound to a value, just use it in begin() and end().
Also set the points to a fixed value. The idea was to "sort" the used
tasks, but this will result in preferring on robot over another, just
because the one has already fetched more bases than the other. We only
want the decision on which robot gets which job based on the location
and by that estimated drive duration.