ASP-Planer: Run in continous mode.
Because the grounding may take a while (which has to be improved!) a
robot may be considered as dead, because we block the CLIPS agent and by
that the callback to set the robots last life signal.

Also that way we can react faster and are not bound by fawkes loop
hooks.