ASP-Planer: Begun with reworking the C++ Part.
We change the procedure from always adding something to the grounder to
just ground once and change the initial state through externals.

For that we have to rework the C++ part and the encoding, but mostly C++
because we now have to save the world state to set the externals in the
correct way.

This commit mainly comments everything we have done out, but the
thread.cpp is reworked.