ASP-Planer: Export robot and zone information.
The robots aliveness, its current task status, and location are
exported. Task status and holding just depend on an empty struct, so
this information is not yet implemented.

Additionally we activate the tasks for zone exploration and the needed
locations on demand.

Updated robots.lp, exploration.lp and facts.lp accordingly.