asp-planner: Change product externals
Products identified by the generating robot will only be used by the
solver wihtin the horizon. The product information at the begin of the
solving are now only identified by an id.

This reduces the externals and by that the needed memory and computation
time. (Time for instance-r2-o1-c2.0-check.lp dropped from ~10 minutes to
~5.) Also the script will be easier to maintain the product overview.