asp-planner: Add goto task, just to go to a location
This enables shorter plans, because a getProduct task will only be
started, when the product is ready for transport. And only at the start
of the task the robot will drive to the location. Now the goto task
kicks in and the robot will drive to the location and start the
getProduct task, when the product is ready, the driveDuration will be 0
in this case. Analog for the other tasks.