asp-planner: Add flush flag and make Python2 compatible
Flushing is needed if we start clingo with --output-debug=text, because
else the output from clingo (stderr) is out of sync with the scripts
output (stdout). But apart from that flushing isn't needed.

Also print("Foo", flush = True) is not compatible with python2, so use a
own print function and flush manually. (Python2 compatibility is not
tested!)