openprs-agent: steps.opf & skills.opf created. A step usually contains a frequently performed elementary action, and any database manipulation required by those actions. Elementary action (ex, skill-calls) are building block of more complex actions.
steps.opf contains the steps collections.
skill.opf contains skills related Ops.

The concept of steps is borrowed from the Clips agent's Steps, but with some adaptation.
Steps here has no sequence nor related to a task, and not activated in
order. They are just some goal
invoked elementary actions, that will be invoked later on from some partial-plan (i.e, Op) that will be picked by the partial planner
(whatever that might be. Could, but does not have to be a Task oriented)