ASP-Planer: Don't store Clingo::Part, store custom GroundRequest.
Because Clingo::Part saves a Clingo::SymbolSpan for the arguments. This
is just a view into an array of Clingo::Symbol, but this array was
temporary in most cases and when we want to ground the memory may hold
already other objects. Fix by copying the symbols into a vector, so we
will manage the lifecycle ourself.