openprs-agent: Steps, Insert-product-into-mps
Insert a product work-peace in a machine's Input for processing of a
production step (RS , DS , CS). Should only be called after correctly
instructing the machine. On succeeding, What ever processing step is
assumed to be complete and the world model is updated. After the call
the Agent is assumed to not hold anything (even if skill failed)
TODO:	make sure if holding or not
	update affected machine facts when created
	update product info when created