openprs-agent: workpiece, predicate template
A WorkPiece serves a central roll in this agent design. Semantically
it represents color information for different materials constituting
a unique  product piece and where it exists (or should exist in case
given as a goal)

ex, (workpiece 23 (base RED)
	(rings yellow orange)(cap GRAY)(place "C-CS1" "output"))

When found  as fact in the database, It represents a unique product
that currently exists in the world, how it looks like and where its
currently located. (e.g, (place "agent" "gripper")
			or (place "MPS-NAME" "SIDE")).

When a workpiece predicate  is as given as a goal. The agent should
be able to figure out how to achieve the described workpiece placing
in the final position given in the goal. The steps required to
achieve such a goal are highly dependent on the worldmodel existing
the moment the goal was posted. Questions like wither there is an
already existing matching workpiece and where it is, effect the
sequence of actions required to achieve the goal.The needed sequence
of steps vary from a simple drive to and gripping action, to more
complex sequences of steps, assembling the goal product from its
constructing parts.
Since the workpiece has very limited "place's" it could exist at,and
a known unique sequence of production steps needed to be completed to
achieve this final piece. An agent should be able to easily reverse
engineer the goal workpiece deconstructing it into how it would look
like and where it would exist during its production sequence.This
continues till either the agent finds the intermediate workpiece with
a matching description in the world,or it the agent knows how to
produce it.