openprs-agent: wp-goals, gripper-to-input
Modifications to wp-gripper-to-input op.

Responds to workpiece goals with (place mps input).
Typically, drives to an mps, instruct it, inster the
workpiece (currently in the gripper) at the mps input

main changes:
-Different sequence of steps depending on machine type
-Restructuring of CS handling
-Instructing the mps before placing the wp (otherwise it
	 will break)

The decision of what to instruct to the ring station is delegated
to another op that should know what instruct based on the final
desired workpiece and the current workpiece the robot is currently
holding.