ASP-Planer: Only deploy consistent plans.
The old way deleted a plan of a robot if there was a problem, but all
other robots may relay on the deleted tasks.

Now if we encounter a problem we do not change any robot plan, but just
start replanning like before. And if there is no problem we deploy the
consistent plan for all robots.