robot-memory-clips: generalize the preconditions for insert-into-cs
There was no use made of the order information and this leads to
inconsitencies when no product-id is set.

This is may not be set by other planners/reasoners than production.clp.
E.g. my ASP-Planer and its executive. This is also useful if you do not
want to commit a product to an order until the cap is mounted, so you
could use the product for another order which has the same prefix.