rcll2016-smt: fix synchronization of symbols
If a fact with a symbol was transmitted it ended up being asserted as a
string, which would fail. To fix this, do not only consider the type of
the incoming data, but also the expected value of the local fact. Only
if both are STRING, assert a string, otherwise assert verbatim.