agent: clear CS even if it is not ready-at-output yet
Most of the time it is convenient to remove an empty base from the CS
direct after inserting it. However, the corresponding production rule
did not fire because the CS was not ready-at-output yet or the agent did
not receive the corresponding message from the refbox yet. Because CS
are fast, we can assume that it well be ready when we arrive at the
output side and thus do not have to wait.