skills:explore_zone: complete rewrite for 2017 game
The exploration changed fundamentally in that there are now 106 zones,
so we cannot explore zone-by-zone any more. Instead, the agent is now
responsible for generating sensible hypotheses for where we might want
to look for MPS'. The explore_zone skill then is only called on a zone
if we have reason to believe that there's actually an MPS there. Making
this assumption allows us to vastly simplify the explore_zone skill.