openprs-agent: Step get-base, generate WP id
On fetch-base Success, Generate the WP id that will used for that base.
This is done to give an Id to a newly generated base only when an agent
holds it. Easier for world WP id sync across agents. Moving the
generations of ids from the reacting on READY_AT_OUTPUT to avoid the
need of syncing a generated workpiece. New workpieces will be given a
zero id till someone uses them.

TODO: Propagate changes to World model