agent: Task Deliver-From-Storage added
Introducing a task retrieval of a C0 from Storage Station and delivering it.
For now, Storage Stations are just another Base-Station with type SS for the
sake of testing. Later on, apply the corrected steps and mps-instructions
specific to SS.
Delivering from storage will be proposed with a pre-defined priority.