openprs-agent: WP, workpiece AtOutput goals[obsolete]
Ops to responding to Workpiece goals that has some mps "output" in
their 'place' flied.

The idea was, Given a Workpiece at some mps output as a goal.The Agent
Reasons about what should be done on that workpiece and preforms it,
by calling Grippe and AtInput operations, with no extra info other than
the goal to be executed.

that is done by finding the earlier version of that workpiece that
exists in the world. Grippe it (if it was not already picked up by
an agent) and processes it at the machine, depending on what was
deduced to be the intended operation. While doing so,achieve-workpiece
facts are updated with the correct processing status. This was meant to
be a more condense goal (workpiece operation) with the purpose of
having a centralized op for calling more basic workpiece operation
(grippe and at input) and monitors their execution.

This is not used anymore (at least for now). Only submitted for
reference.