openprs-agent: Prod_multi, Tasks Ops related to C0
Production of C0 tasks responding to goals in the form
(! Task-XXX $agent). Upon calling the task.
Each task starts by a waits for its perconditions to be
satisfied (sleeping that goal till so happens). When finally
woken up the Task op posts the correct achieve-workpiece goal
(which in turn blocks till the achieve-workpiece facts reach
a final state. ie,SUCCEEDED or FAILED)

Task--Retrieve-cap
Task--Clear-Cap-station
Task--Mount-Cap-C0
Task--Deliver

This is just the skillton of whats going to be a more elaborate way of
executing the tasks.