openprs-agent: prod_multi,2 agents Produce C0
After picking a goal in-mind (by an explicit call for testing).
Intend all Productions Tasks required to produce a c0 as roots.

At this point there is no way of knowing which Tasks succeeded or
failed or stuck. The tasks take care of their own execution once
the preconditions that it waited for are valid. Task have do not
care about how the conditions where satisfied.