openprs-agent: Synced Assertion of WP as Achievment
Asserting a WP done by asserting an Achieve-workpiece with a
SUCCEEDED status, including the agent name and the WP info.

This ensures the workpiece is sent to the rest of the agents.

Specially important in Steps and Workpiece-operations since till this
point they only updated the agents own world model.